save theme preference
