+++ android/app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:label="NekoList"
         android:name="${applicationName}"
         android:usesCleartextTraffic="true"
-        android:icon="@mipmap/launcher_icon">
+        android:icon="@mipmap/ic_launcher">
         <activity
             android:name=".MainActivity"
             android:exported="true"
