+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:neko_list/theme/theme_provider.dart';
 import 'package:provider/provider.dart';
 import 'screens/auth/login.dart';
-import 'helpers/secure_storage.dart';
+import 'helpers/storage.dart';
 import 'screens/homepage.dart';
 
 void main() {
