+++ pubspec.yaml
@@ -99,6 +99,7 @@ flutter:
   # see https://flutter.dev/custom-fonts/#from-packages
 
 flutter_icons:
-  android: "launcher_icon"
+  android: true
   ios: true
-  image_path: "assets/icons/launcher_icon.png"
+  remove_alpha_ios: true
+  image_path: "assets/icons/logo_catgirl.png"
